Release builder fixes to beta vs non-beta handling (#1528)
Allows any permutation of a beta/non-beta stub universe => beta/non-beta release.
Automatically updates the package name and version to reflect beta state as needed. This handling is now done up-front to make things a bit easier to follow. Complains if a -beta version is provided with beta mode disabled.